Fix most of the following classes of warnings in the ToolTalk libraries:
- Const strings referenced by non-const variables.
- Incorrect format specifers for printing addresses
- Unused variables
- Signed comparison to unsigned

Also fix an incorrect enumeration value in a switch statement.

+++ cde/lib/tt/lib/db/tt_db_key.C
@@ -44,7 +44,7 @@
 
 _Tt_db_key::_Tt_db_key (short version_number)
 {
-  static long last_time_sec = 0;
+  static unsigned int last_time_sec = 0;
   static long counter = 0;
 
   key.version  = version_number;
